% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiledb_vcf.R
\name{tiledb_vcf_available}
\alias{tiledb_vcf_available}
\title{Check TileDB-VCF Availability}
\usage{
tiledb_vcf_available()
}
\value{
A logical value indicating if TileDB-VCF is available.
}
\description{
Checks if the TileDB-VCF library is available and functional.
}
\examples{
tiledb_vcf_available()
}
