% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_interface.R
\name{tiledb_vcf_cli}
\alias{tiledb_vcf_cli}
\title{TileDB-VCF Command Line Interface}
\description{
These functions provide R interfaces to the TileDB-VCF command line utility.
They use the \code{processx} package to execute the TileDB-VCF binary that is
installed with this package.
}
\details{
The TileDB-VCF CLI provides commands for:
\itemize{
\item \code{create}: Creating empty TileDB-VCF datasets
\item \code{store}: Ingesting VCF files into datasets
\item \code{export}: Exporting data from datasets
\item \code{list}: Listing samples in datasets
\item \code{stat}: Getting dataset statistics
\item \code{delete}: Deleting samples from datasets
}
}
