% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_interface.R
\name{tiledb_vcf_cli_path}
\alias{tiledb_vcf_cli_path}
\title{Get TileDB-VCF CLI Path}
\usage{
tiledb_vcf_cli_path()
}
\value{
Character string with the path to the TileDB-VCF binary.
}
\description{
Returns the path to the TileDB-VCF command line binary installed with this package.
}
\examples{
\dontrun{
tiledb_vcf_cli_path()
}
}
