% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_interface.R
\name{tiledb_vcf_cli_version}
\alias{tiledb_vcf_cli_version}
\title{Get TileDB-VCF CLI Version}
\usage{
tiledb_vcf_cli_version(echo = FALSE)
}
\arguments{
\item{echo}{Logical, whether to echo output to console (default: FALSE)}
}
\value{
Character string with version information
}
\description{
Get the version of the TileDB-VCF command line utility.
}
\examples{
\dontrun{
tiledb_vcf_cli_version()
}
}
