% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiledb_vcf.R
\name{tiledb_vcf_include_headers}
\alias{tiledb_vcf_include_headers}
\title{Get TileDB-VCF Include Headers
Returns the include headers required to compile code that uses the TileDB-VCF library.}
\usage{
tiledb_vcf_include_headers()
}
\value{
A character string containing the include headers.
}
\description{
Get TileDB-VCF Include Headers
Returns the include headers required to compile code that uses the TileDB-VCF library.
}
