% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiledb_vcf.R
\name{tiledb_vcf_include_path}
\alias{tiledb_vcf_include_path}
\title{Get TileDB-VCF Include Path
Returns the include path required to compile code that uses the TileDB-VCF library.}
\usage{
tiledb_vcf_include_path()
}
\value{
A character string containing the include path.
}
\description{
Get TileDB-VCF Include Path
Returns the include path required to compile code that uses the TileDB-VCF library.
}
