% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiledb_vcf.R
\name{tiledb_vcf_library_items}
\alias{tiledb_vcf_library_items}
\title{Get TileDB-VCF Library Items
Returns the full path to the library items required to link against the TileDB-VCF libraries}
\usage{
tiledb_vcf_library_items()
}
\value{
A character string containing the library items.
}
\description{
Get TileDB-VCF Library Items
Returns the full path to the library items required to link against the TileDB-VCF libraries
}
