% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiledb_vcf.R
\name{tiledb_vcf_library_path}
\alias{tiledb_vcf_library_path}
\title{Get the TileDB-VCF Library Path}
\usage{
tiledb_vcf_library_path()
}
\value{
A character string containing the TileDB-VCF library path.
}
\description{
Returns the path to the TileDB-VCF libraries installed with the Rlibtiledbvcf package.
}
