% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiledb_vcf.R
\name{tiledb_vcf_version}
\alias{tiledb_vcf_version}
\title{Get TileDB-VCF Version}
\usage{
tiledb_vcf_version()
}
\value{
A character string containing the TileDB-VCF version.
}
\description{
Returns the version of the TileDB-VCF library.
}
\examples{
tiledb_vcf_version()
}
