% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiledb_vcf.R
\name{tiledb_vcf_pkg_cflags}
\alias{tiledb_vcf_pkg_cflags}
\title{Get TileDB-VCF Compilation Flags
Returns the compilation flags required to compile code that uses the TileDB-VCF library.}
\usage{
tiledb_vcf_pkg_cflags()
}
\value{
A character string containing the compilation flags.
}
\description{
Get TileDB-VCF Compilation Flags
Returns the compilation flags required to compile code that uses the TileDB-VCF library.
}
\examples{
\dontrun{
# Example of using tiledb_vcf_pkg_cflags in a compilation command
system(paste("R CMD SHLIB -o my_program.so my_program.c", tiledb_vcf_pkg_lflags(), tiledb_vcf_pkg_cflags()))
}
}
